cp spark code
