+++ src/main/java/org/apache/parquet/hadoop/VectorizedOapRecordReader.java
@@ -29,7 +29,7 @@ import org.apache.parquet.schema.Type;
 
 import org.apache.spark.memory.MemoryMode;
 import org.apache.spark.sql.catalyst.InternalRow;
-import org.apache.spark.sql.execution.datasources.parquet.VectorizedColumnReaderWrapper;
+import org.apache.spark.sql.execution.datasources.parquet.OapVectorizedColumnReader;
 import org.apache.spark.sql.execution.vectorized.ColumnVectorUtils;
 import org.apache.spark.sql.execution.vectorized.ColumnarBatch;
 import org.apache.spark.sql.types.StructField;
@@ -54,7 +54,7 @@ public class VectorizedOapRecordReader extends SpecificOapRecordReaderBase<Objec
      * From VectorizedParquetRecordReader, change private to protected,
      * wrapper VectorizedColumnReader.
      */
-    protected VectorizedColumnReaderWrapper[] columnReaders;
+    protected OapVectorizedColumnReader[] columnReaders;
 
     /**
      * The number of rows that have been returned.
@@ -364,10 +364,10 @@ public class VectorizedOapRecordReader extends SpecificOapRecordReaderBase<Objec
         + rowsReturned + " out of " + totalRowCount);
       }
       List<ColumnDescriptor> columns = requestedSchema.getColumns();
-      columnReaders = new VectorizedColumnReaderWrapper[columns.size()];
+      columnReaders = new OapVectorizedColumnReader[columns.size()];
       for (int i = 0; i < columns.size(); ++i) {
         if (missingColumns[i]) continue;
-        columnReaders[i] = new VectorizedColumnReaderWrapper(
+        columnReaders[i] = new OapVectorizedColumnReader(
           columns.get(i), pages.getPageReader(columns.get(i)));
       }
       totalCountLoadedSoFar += pages.getRowCount();
