--- src/main/java/org/apache/spark/sql/execution/datasources/parquet/VectorizedColumnReaderWrapper.java
@@ -1,47 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.spark.sql.execution.datasources.parquet;
-
-import java.io.IOException;
-
-import org.apache.parquet.column.ColumnDescriptor;
-import org.apache.parquet.column.page.PageReader;
-import org.apache.spark.sql.execution.vectorized.ColumnVector;
-
-/**
- * VectorizedColumnReaderWrapper let readBatch method can be used
- * outside "org.apache.spark.sql.execution.datasources.parquet" package
- */
-public class VectorizedColumnReaderWrapper {
-
-    private OapVectorizedColumnReader reader;
-
-    public VectorizedColumnReaderWrapper(
-        ColumnDescriptor descriptor, PageReader pageReader) throws IOException {
-      this.reader = new OapVectorizedColumnReader(descriptor, pageReader);
-    }
-
-    public void readBatch(int total, ColumnVector column) throws IOException {
-      reader.readBatch(total, column);
-    }
-
-    public void skipBatch(int total, ColumnVector column) throws IOException {
-      reader.skipBatch(total, column);
-    }
-
-}
