+++ src/main/scala/org/apache/spark/sql/execution/datasources/oap/io/ParquetFiberDataLoader.scala
@@ -28,7 +28,7 @@ import org.apache.parquet.schema.{MessageType, Type}
 import org.apache.spark.memory.MemoryMode
 import org.apache.spark.sql.execution.datasources.OapException
 import org.apache.spark.sql.execution.datasources.oap.filecache.FiberCache
-import org.apache.spark.sql.execution.datasources.parquet.{ParquetReadSupportWrapper, VectorizedColumnReader, VectorizedColumnReaderWrapper}
+import org.apache.spark.sql.execution.datasources.parquet.{OapVectorizedColumnReader, ParquetReadSupportWrapper}
 import org.apache.spark.sql.execution.vectorized.{ColumnVector, OnHeapColumnVector, OnHeapColumnVectorFiber}
 import org.apache.spark.sql.oap.OapRuntime
 import org.apache.spark.sql.types._
@@ -78,8 +78,8 @@ private[oap] case class ParquetFiberDataLoader(
         val blockMetaData = footer.getBlocks.get(blockId)
         val fiberData = reader.readFiberData(blockMetaData, columnDescriptor)
         val columnReader =
-          new VectorizedColumnReaderWrapper(
-            columnDescriptor, fiberData.getPageReader(columnDescriptor));
+          new OapVectorizedColumnReader(
+            columnDescriptor, fiberData.getPageReader(columnDescriptor))
         columnReader.readBatch(rowCount, vector)
       }
