+++ src/test/scala/org/apache/spark/sql/execution/datasources/oap/io/ParquetDataFileSuite.scala
@@ -41,7 +41,7 @@ import org.apache.spark.SparkFunSuite
 import org.apache.spark.internal.Logging
 import org.apache.spark.memory.MemoryMode
 import org.apache.spark.sql.execution.datasources.oap.filecache.FiberCache
-import org.apache.spark.sql.execution.datasources.parquet.{ParquetReadSupportWrapper, VectorizedColumnReader, VectorizedColumnReaderWrapper}
+import org.apache.spark.sql.execution.datasources.parquet.{OapVectorizedColumnReader, ParquetReadSupportWrapper}
 import org.apache.spark.sql.execution.vectorized.{ColumnarBatch, ColumnVector}
 import org.apache.spark.sql.internal.SQLConf
 import org.apache.spark.sql.internal.oap.OapConf
@@ -536,7 +536,7 @@ class ParquetFiberDataReaderSuite extends ParquetDataFileSuite {
     val columnDescriptor = parquetSchema.getColumns.get(0)
     val fiberData = reader.readFiberData(blockMetaData, columnDescriptor)
     val columnReader =
-      new VectorizedColumnReaderWrapper(
+      new OapVectorizedColumnReader(
         columnDescriptor, fiberData.getPageReader(columnDescriptor))
     columnReader.readBatch(rowCount, vector)
     for (i <- 0 until rowCount) {
@@ -550,7 +550,6 @@ class ParquetFiberDataReaderSuite extends ParquetDataFileSuite {
       new Path(fileName), meta.footer.toParquetMetadata)
     val footer = reader.getFooter
     val rowCount = footer.getBlocks.get(0).getRowCount.toInt
-    val vector = ColumnVector.allocate(rowCount, IntegerType, MemoryMode.ON_HEAP)
     val blockMetaData = footer.getBlocks.get(0)
     val columnDescriptor = new ColumnDescriptor(Array(s"${fileName}_temp"), INT32, 0, 0)
     val exception = intercept[IOException] {
